+++ src/main/java/com/ll/jsp/board/boundedContext/article/service/ArticleService.java
@@ -22,7 +22,7 @@ public class ArticleService {
         return articleRepository.save(title, content);
     }
 
-    public Article findById(int id) {
+    public Article findById(long id) {
         return articleRepository.findById(id);
     }
 }
